+++ contrib/go/src/python/pants/contrib/go/subsystems/go_distribution.py
@@ -10,7 +10,7 @@ from builtins import str
 from pants.base.workunit import WorkUnit, WorkUnitLabel
 from pants.binaries.binary_tool import NativeTool
 from pants.binaries.binary_util import BinaryToolUrlGenerator
-from pants.util.collections_backport import OrderedDict, namedtuple
+from pants.util.collections_abc_backport import OrderedDict, namedtuple
 from pants.util.memo import memoized_property
 from pants.util.process_handler import subprocess
