+++ contrib/python/src/python/pants/contrib/python/checks/checker/BUILD
@@ -10,7 +10,7 @@ python_library(
     }
   ),
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python:pycodestyle',
     '3rdparty/python:pyflakes',
