+++ src/python/pants/backend/codegen/protobuf/java/BUILD
@@ -3,7 +3,7 @@
 
 python_library(
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     'src/python/pants/backend/codegen/protobuf/subsystems',
