+++ src/python/pants/backend/codegen/protobuf/java/protobuf_gen.py
@@ -21,7 +21,7 @@ from pants.base.workunit import WorkUnitLabel
 from pants.build_graph.address import Address
 from pants.fs.archive import ZIP
 from pants.task.simple_codegen_task import SimpleCodegenTask
-from pants.util.collections_backport import OrderedDict
+from pants.util.collections_abc_backport import OrderedDict
 from pants.util.process_handler import subprocess
