+++ src/python/pants/backend/jvm/BUILD
@@ -45,7 +45,7 @@ python_library(
   name='ivy_utils',
   sources=['ivy_utils.py'],
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     '3rdparty/python:future',
     '3rdparty/python:six',
