+++ src/python/pants/backend/jvm/tasks/BUILD
@@ -174,7 +174,7 @@ python_library(
   name = 'classpath_util',
   sources = ['classpath_util.py'],
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     ':classpath_entry',
     '3rdparty/python:future',
     '3rdparty/python/twitter/commons:twitter.common.collections',
@@ -323,7 +323,7 @@ python_library(
   name = 'jar_publish',
   sources = ['jar_publish.py'],
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     ':jar_publish_resources',
@@ -600,7 +600,7 @@ python_library(
   name = 'properties',
   sources = ['properties.py'],
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python:six',
   ],
