+++ src/python/pants/backend/jvm/tasks/classpath_util.py
@@ -10,7 +10,7 @@ from builtins import filter, object
 from twitter.common.collections import OrderedSet
 
 from pants.backend.jvm.tasks.classpath_entry import ClasspathEntry
-from pants.util.collections_backport import OrderedDict
+from pants.util.collections_abc_backport import OrderedDict
 from pants.util.contextutil import open_zip
 from pants.util.dirutil import fast_relpath, safe_walk
 from pants.util.strutil import ensure_text
