+++ src/python/pants/backend/jvm/tasks/jar_publish.py
@@ -35,7 +35,7 @@ from pants.ivy.bootstrapper import Bootstrapper
 from pants.ivy.ivy import Ivy
 from pants.task.scm_publish_mixin import Namedver, ScmPublishMixin, Semver
 from pants.task.target_restriction_mixins import HasTransitiveOptionMixin, TransitiveOptionRegistrar
-from pants.util.collections_backport import OrderedDict, defaultdict, namedtuple
+from pants.util.collections_abc_backport import OrderedDict, defaultdict, namedtuple
 from pants.util.dirutil import safe_mkdir, safe_open, safe_rmtree
 from pants.util.strutil import ensure_text
