+++ src/python/pants/backend/python/tasks/BUILD
@@ -3,7 +3,7 @@
 
 python_library(
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:configparser',
     '3rdparty/python:future',
     '3rdparty/python:pex',
