+++ src/python/pants/backend/python/tasks/setup_py.py
@@ -34,7 +34,7 @@ from pants.build_graph.address_lookup_error import AddressLookupError
 from pants.build_graph.build_graph import sort_targets
 from pants.build_graph.resources import Resources
 from pants.task.task import Task
-from pants.util.collections_backport import (Iterable, Mapping, MutableSequence, OrderedDict, Set,
+from pants.util.collections_abc_backport import (Iterable, Mapping, MutableSequence, OrderedDict, Set,
                                              defaultdict)
 from pants.util.contextutil import temporary_file
 from pants.util.dirutil import safe_concurrent_creation, safe_rmtree, safe_walk
