+++ src/python/pants/build_graph/BUILD
@@ -4,7 +4,7 @@
 
 python_library(
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     '3rdparty/python:future',
     '3rdparty/python:six',
