+++ src/python/pants/build_graph/build_graph.py
@@ -15,7 +15,7 @@ from pants.build_graph.address import Address
 from pants.build_graph.address_lookup_error import AddressLookupError
 from pants.build_graph.injectables_mixin import InjectablesMixin
 from pants.build_graph.target import Target
-from pants.util.collections_backport import OrderedDict, defaultdict, deque
+from pants.util.collections_abc_backport import OrderedDict, defaultdict, deque
 from pants.util.meta import AbstractClass
