+++ src/python/pants/cache/BUILD
@@ -3,7 +3,7 @@
 
 python_library(
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python:requests',
     '3rdparty/python:pyopenssl',
