+++ src/python/pants/cache/pinger.py
@@ -12,7 +12,7 @@ import requests
 from future.moves.urllib.parse import urlparse
 
 from pants.cache.artifact_cache import ArtifactCacheError
-from pants.util.collections_backport import Counter, deque
+from pants.util.collections_abc_backport import Counter, deque
 from pants.util.contextutil import Timer
 from pants.util.memo import memoized_method
