+++ src/python/pants/engine/BUILD
@@ -13,7 +13,7 @@ python_library(
   name = 'legacy_engine',
   sources = ['legacy_engine.py', 'round_engine.py', 'round_manager.py'],
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     'src/python/pants/base:exceptions',
@@ -27,7 +27,7 @@ python_library(
   name='addressable',
   sources=['addressable.py'],
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     ':objects',
     'src/python/pants/build_graph',
@@ -40,7 +40,7 @@ python_library(
   name='struct',
   sources=['struct.py'],
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     ':objects',
     'src/python/pants/util:objects',
@@ -67,7 +67,7 @@ python_library(
   name='build_files',
   sources=['build_files.py'],
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python:six',
     ':addressable',
@@ -89,7 +89,7 @@ python_library(
   name='mapper',
   sources=['mapper.py'],
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     ':objects',
     ':parser',
     'src/python/pants/build_graph',
@@ -146,7 +146,7 @@ python_library(
   name='rules',
   sources=['rules.py'],
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     ':selectors',
