+++ src/python/pants/engine/struct.py
@@ -8,7 +8,7 @@ from future.utils import binary_type, text_type
 
 from pants.engine.addressable import addressable, addressable_list
 from pants.engine.objects import Serializable, SerializableFactory, Validatable, ValidationError
-from pants.util.collections_backport import MutableMapping, MutableSequence
+from pants.util.collections_abc_backport import MutableMapping, MutableSequence
 from pants.util.objects import SubclassesOf, SuperclassesOf
