+++ src/python/pants/fs/archive.py
@@ -12,7 +12,7 @@ from zipfile import ZIP_DEFLATED
 
 from future.utils import PY2
 
-from pants.util.collections_backport import OrderedDict
+from pants.util.collections_abc_backport import OrderedDict
 from pants.util.contextutil import open_tar, open_zip, temporary_dir
 from pants.util.dirutil import is_executable, safe_concurrent_rename, safe_walk
 from pants.util.meta import AbstractClass
