+++ src/python/pants/help/BUILD
@@ -3,7 +3,7 @@
 
 python_library(
   dependencies=[
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
