+++ src/python/pants/help/build_dictionary_info_extracter.py
@@ -11,7 +11,7 @@ from builtins import object, range
 
 from pants.base.exceptions import TaskError
 from pants.build_graph.target import Target
-from pants.util.collections_backport import OrderedDict, namedtuple
+from pants.util.collections_abc_backport import OrderedDict, namedtuple
 from pants.util.memo import memoized_method
