+++ src/python/pants/option/options_fingerprinter.py
@@ -14,7 +14,7 @@ from pants.base.build_environment import get_buildroot
 from pants.base.hash_utils import stable_json_hash
 from pants.option.custom_types import (UnsetBool, dict_with_files_option, dir_option, file_option,
                                        target_option)
-from pants.util.collections_backport import Iterable, Mapping
+from pants.util.collections_abc_backport import Iterable, Mapping
 
 
 class Encoder(json.JSONEncoder):
