+++ src/python/pants/util/objects.py
@@ -11,7 +11,7 @@ from builtins import object, zip
 from future.utils import PY2
 from twitter.common.collections import OrderedSet
 
-from pants.util.collections_backport import OrderedDict, namedtuple
+from pants.util.collections_abc_backport import OrderedDict, namedtuple
 from pants.util.memo import memoized, memoized_classproperty
 from pants.util.meta import AbstractClass
