+++ tests/python/pants_test/backend/jvm/tasks/BUILD
@@ -181,7 +181,7 @@ python_tests(
   name = 'classpath_util',
   sources = ['test_classpath_util.py'],
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     'src/python/pants/java/jar',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/tasks:classpath_products',
@@ -668,7 +668,7 @@ python_tests(
   name = 'properties',
   sources = ['test_properties.py'],
   dependencies = [
-    'src/python/pants/util:collections_backport',
+    'src/python/pants/util:collections_abc_backport',
     '3rdparty/python:future',
     'src/python/pants/backend/jvm/tasks:properties',
   ]
